+++ src/GEmojiSharp.AspNetCore/GEmojiSharp.AspNetCore.csproj
